Remove storing of handicap
Fixes #20

Handicaps can be calculated quickly based on the average of the user.  Remove the handicap field and make the average a first tier member of the bowler definition object.

Handicaps make more sense for the bowler instance objects because that is locked in time and is not calculated as the weeks move forward.